Create a reusable function for committing changes
Committing changes is not as simple as it may seem. It involves two steps both
of which may fail and we need to properly report such failure. And we are going
to need to do this in every function from this plugin. Thus it will be useful to
have a separate, shared function for this.