Implement the support for BD_PART_TYPE_REQ_NEXT
If the requested type of a to-be-created partition is not specified, we need to
decide on our own based on the state of the disk. If there already is an
extended partition, we create a logical parititon, otherwise if only the last
primary or extended partition can be created, we create an extended partition
and a logical partition in it. If it's no special case, we just create a primary
partition.