Take new size for an XFS FS in a number of blocks
Otherwise we would need both the device and its mount point. Let's leave it up
to the caller to get the block size from xfs_get_info() output and tell us the
requested number of blocks.

Also use LVM for tests instead of raw DM. This should save us from a lot of pain
with things taking some random amount of time to really process and finish with
raw DM.