Specify tolerance for partition size
If the caller of the partition creation and
resize functions is not aware of size reductions
due to alignment it should still be able to give
the theoretical free space as size argument.
This should be possible by a 3 MiB tolerance which
the final partition is allowed to be smaller.